refactor: split the autostart tests into their own file
Adding the uninstall tests pushed internal/autostart/mod.rs to 527 code
lines, over the 500-line hard limit, and line-limit.yml rightly failed
the build.

Moved the test module into autostart/tests.rs, matching how the
quadlet-mode tests already live in quadlet/tests.rs. Both files land well
under the 300-line soft limit (273 and 253), and the `unix` gate is
unchanged. No test bodies were touched — 1260/1260 still pass.

Part of #1080
